# lib/api-zod/src/generated/types/synonymAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UsageExample } from "./usageExample";
import type { WordNuance } from "./wordNuance";

export interface SynonymAnalysisResult {
  sharedMeaning: string;
  nuances: WordNuance[];
  memoryTips: string;
  usageExamples: UsageExample[];
}