# lib/api-zod/src/generated/types/wordMeaningResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WordMeaningResult {
  word: string;
  meaning: string;
  nuance: string;
  usageHints: string;
  similarWords: string;
  memoryTip: string;
}